requirements.txt
ansible==5.1.0
ansible-core==2.12.1
ansible-modules-hashivault==4.6.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
hvac==0.11.2
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
Pillow==8.4.0
pycparser==2.21
pyhcl==0.4.4
pyparsing==3.0.6
python-gitlab==2.10.1
PyYAML==6.0
requests==2.26.0
requests-toolbelt==0.9.1
resolvelib==0.5.4
six==1.16.0
urllib3==1.26.7
